requirements.txt
python-dotenv
streamlit
langchain-openai
langchain-community
chromadb
pypdf


